requirements.txt
torch==2.3.0
transformers>=4.36.2
datasets>=2.14.3
accelerate>=0.21.0
peft>=0.7.0
trl>=0.7.6
gradio>=3.38.0,<4.0.0
scipy
einops
sentencepiece
protobuf
jieba
rouge-chinese
nltk
uvicorn
pydantic
fastapi
sse-starlette
matplotlib
elasticsearch==8.12.1
langchain==0.1.9
pandas==1.4.4
openpyxl
sentence_transformers
transformers_stream_generator
tiktoken